Add files via upload
This is my Car_ShowRoom_Management_System project. In this project, I'm using Java with JDBC to Java with a database. In this project, Im using MySql to store my data in a database. This project is divided into 3 parts: Admin, Client, and Normal-User.
In the Admin mode program run with full permission on the database, Admin can delete, add, and check all the data from the database.
In client mode program can have only partial data read/write permission, in partial program can only read/write data under that client.
In Normal-User mode no access to the database user can just access the catalog page to see the price with details.